Mise en place de pattern HTML pour la validation des formulaires de creation du site

+++ amis-de-escalade/couche-webapp/src/main/java/org/couche/webapp/servlets/CreationSite.java
@@ -90,7 +90,7 @@ public class CreationSite extends HttpServlet {
 		// Création des secteurs à vide
 		SecteurService secteurService = new SecteurService();
 		
-		for(int i = 0; i < Long.parseLong(nombreSecteurs); i++) {
+		for(int i = 0; i < Integer.parseInt(nombreSecteurs); i++) {
 			
 			secteurService.createSecteur("", i + 1, "", site);			
 		}
